backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
pydantic==2.9.0
pydantic[email]==2.9.0
mangum==0.18.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9